+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   see [HMKitFleet.sendCommand](hmkit-fleet/src/main/kotlin/HMKitFleet.kt)
   and  [TelematicsResponse](hmkit-fleet/src/main/kotlin/network/Response.kt)
 
+## [0.5.2]
+- add getClearanceStatus:vin method
+- undeprecate ClearanceStatus.Status.ERROR
+
 ## [0.5.1]
 
 - add ford brand for fleet vehicles